move bedcropdupe into antidupe module
